feat: add useCommentLikeManager
