+++ packages/client/src/hooks/queries/comment/useLikeComment.ts
@@ -1,0 +1,9 @@
+import CommentAPI from '~/api/comment';
+import { useMutation } from '@tanstack/react-query';
+import { UseMutationOptionsOf } from '~/hooks/queries/types';
+
+const useLikeComment = (options: UseMutationOptionsOf<typeof CommentAPI.likeComment> = {}) => {
+  return useMutation(CommentAPI.likeComment, options);
+};
+
+export default useLikeComment;
